[feat/#78]: 컴포넌트 분리
